requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements-dev.in
#
astor==0.8.1
    # via flake8-qgis
atomicwrites==1.4.0
    # via
    #   -r requirements-dev.in
    #   pytest
attrs==23.1.0
    # via pytest
black==23.7.0
    # via -r requirements-dev.in
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage==7.3.2
    # via pytest-cov
distlib==0.3.7
    # via virtualenv
filelock==3.12.4
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements-dev.in
    #   flake8-pyproject
    #   flake8-qgis
flake8-pyproject==1.2.3
    # via -r requirements-dev.in
flake8-qgis==1.0.0
    # via -r requirements-dev.in
identify==2.5.30
    # via pre-commit
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.4.1
    # via -r requirements-dev.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via
    #   -c requirements.txt
    #   black
    #   pytest
pathspec==0.11.2
    # via black
platformdirs==3.11.0
    # via
    #   black
    #   virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.4.0
    # via -r requirements-dev.in
py==1.11.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyqt5-stubs==5.15.6.0
    # via -r requirements-dev.in
pytest==6.2.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-mock
    #   pytest-qgis
    #   pytest-qt
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-mock==3.7.0
    # via -r requirements-dev.in
pytest-qgis==1.3.1
    # via -r requirements-dev.in
pytest-qt==3.3.0
    # via -r requirements-dev.in
pyyaml==6.0.1
    # via pre-commit
ruff==0.0.292
    # via -r requirements-dev.in
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==2.0.1
    # via
    #   -c requirements.txt
    #   black
    #   flake8-pyproject
    #   mypy
typing-extensions==4.8.0
    # via
    #   black
    #   mypy
virtualenv==20.24.5
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
